created separate view controller for email login
