+++ SocialApp/SignInVC.swift
@@ -14,15 +14,13 @@ import SwiftKeychainWrapper
 
 // /Users/chasemcelroy/Development/Tutorials/SocialApp/SocialApp/SignInVC.swift:38:72: Use 'String(describing:)' to silence this warning
 
-class SignInVC: UIViewController, UITextFieldDelegate {
+class SignInVC: UIViewController {
 
-    @IBOutlet weak var emailField: FancyField!
-    @IBOutlet weak var pwdField: FancyField!
+
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.emailField.delegate = self
-        self.pwdField.delegate = self
+
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -75,33 +73,6 @@ class SignInVC: UIViewController, UITextFieldDelegate {
         })
     }
     
-    @IBAction func signInTapped(_ sender: Any) {
-        if let email = emailField.text, let pwd = pwdField.text {
-            FIRAuth.auth()?.signIn(withEmail: email, password: pwd, completion: { (user, error) in
-                if error == nil {
-                    print("CHASE: EMAIL User authenticated with Firebase")
-                    if let user = user {
-                        let userData = [PROVIDER_DB_STRING: user.providerID,
-                                        EMAIL_DB_STRING: email]
-                        self.completeSignIn(id: user.uid, userData: userData)
-                    }
-                } else {
-                    FIRAuth.auth()?.createUser(withEmail: email, password: pwd, completion: { (user, error) in
-                        if error != nil {
-                            print("CHASE: unable to authenticate with Firebase user email")
-                        } else {
-                            print("CHASE: Succesffully authentitcated with Firebase email")
-                            if let user = user {
-                                let userData = ["provider": user.providerID,
-                                                EMAIL_DB_STRING: email]
-                                self.completeSignIn(id: user.uid, userData: userData)
-                            }
-                        }
-                    })
-                }
-            })
-        }
-    }
     
     func completeSignIn(id: String, userData: Dictionary<String, String>) {
         
@@ -121,40 +92,9 @@ class SignInVC: UIViewController, UITextFieldDelegate {
         })
     }
     
-    
-    // MARK: KEYBOARD FUNCTIONS
-    // Move View
-    func moveTextField(textField: UITextField, moveDistance: Int, up: Bool) {
-        let moveDuration = 0.3
-        let movement: CGFloat = CGFloat(up ? moveDistance : -moveDistance)
-        
-        UIView.beginAnimations("animateTextField", context: nil)
-        UIView.setAnimationBeginsFromCurrentState(true)
-        UIView.setAnimationDuration(moveDuration)
-        self.view.frame = self.view.frame.offsetBy(dx: 0, dy: movement)
-        UIView.commitAnimations()
+    @IBAction func emailBtnTapped(_ sender: Any) {
+        performSegue(withIdentifier: "emailSegue", sender: nil)
     }
     
-    // Keyboard shows
-    func textFieldDidBeginEditing(_ textField: UITextField) {
-        moveTextField(textField: textField, moveDistance: -250, up: true)
-    }
-    
-    // Keyboard is hidden
-    func textFieldDidEndEditing(_ textField: UITextField) {
-        moveTextField(textField: textField, moveDistance: -250, up: false)
-    }
-    
-    //presses return key
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        emailField.resignFirstResponder()
-        pwdField.resignFirstResponder()
-        return true
-    }
-    
-    // Hide keyboard when user touches outside keyboard
-    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        self.view.endEditing(true)
-    }
-}
+   }
